Change the release info display
We used to display it in place of the version (in the sidebar under the
logo) when it was defined in `conf.py`. This value is however intended
at holding unique Git-based revision, hence doesn't fit the previous
usage.
Instead, we only use the version under the logo, and display the release
at the bottom of the sidebar if available.