Fix the `copy-block` directive vertical scroll
On mac screen (using chrome and firefox) when hovering the copy button the copy label was appearing a bit too down due to the padding which was making a vertical scroll appearing next to the copy button which change it's position making a bit more difficult to use. Adding a margin-top solves the issue.